+++ src/index.html
@@ -11,6 +11,7 @@
   <link rel="stylesheet" href="<%=htmlWebpackPlugin.files.chunks.main.css%>">
   <link rel="icon" href="./image/favicon/N.png">
   <title>Главная</title>
+
 </head>
 
 <body class="page">
@@ -106,7 +107,7 @@
     </h2>
     <span class="saved__key-words none">
       По ключевым словам:<b class="saved__key-words-bold"> </b>
-  </span>
+    </span>
     <p class="header__content-description">
       Находите самые свежие статьи на любую тему и сохраняйте в своём личном кабинете.
     </p>
