+++ src/news.html
@@ -11,6 +11,7 @@
     <link rel="stylesheet" href="<%=htmlWebpackPlugin.files.chunks.main.css %>">
     <link rel="icon" href="./image/favicon/N.png">
     <title>Сохраненные статьи</title>
+    
 </head>
 
 <body class="page">
@@ -270,6 +271,9 @@
                     Идет поиск новостей...
                 </h3>
             </div>
+            <button class="header__auth-button header__auth-button_view-more button__effect none">
+                Показать еще
+              </button>
             <span class="saved__key-words">
                 По ключевым словам:<b class="saved__key-words-bold"> </b>
             </span>
